fix(cli): clean up connection commands
