+++ packages/cli/src/notion-page-picker-tui.tsx
@@ -9,8 +9,8 @@ import {
   visibleNodeIds,
   type PickerCommand,
   type PickerState,
-} from './connection-notion-tree.js';
-import type { KtxCliIo } from '../index.js';
+} from './notion-page-picker-tree.js';
+import type { KtxCliIo } from './cli-runtime.js';
 
 const COLOR_THEME = {
   text: 'white',
